+++ app/src/main/java/it/gym/exception/ConflictException.java
@@ -3,9 +3,10 @@ package it.gym.exception;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class TrainingSessionNotFound extends RuntimeException {
-    public TrainingSessionNotFound(String message) {
+@ResponseStatus(HttpStatus.CONFLICT)
+public class ConflictException extends RuntimeException {
+
+    public ConflictException(String message) {
         super(message);
     }
 }