+++ app/src/main/java/it/gym/exception/InternalServerException.java
@@ -3,10 +3,9 @@ package it.gym.exception;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class ExpiredTokenException extends RuntimeException {
-
-    public ExpiredTokenException(String message) {
+@ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+public class InternalServerException extends RuntimeException {
+    public InternalServerException(String message) {
         super(message);
     }
 }