+++ app/src/main/java/it/gym/handler/CustomizedResponseEntityExceptionHandler.java
@@ -40,157 +40,44 @@ public class CustomizedResponseEntityExceptionHandler extends ResponseEntityExce
         return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
     }
 
-    @ExceptionHandler(InternalReservationException.class)
-    public final ResponseEntity<ErrorDetails> handle(InternalReservationException ex,
+    @ExceptionHandler(InternalServerException.class)
+    public final ResponseEntity<ErrorDetails> handle(InternalServerException ex,
                                                      WebRequest request) {
         ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
                 request.getDescription(false));
         return new ResponseEntity<>(errorDetails, HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
-    @ExceptionHandler(ExpiredTokenException.class)
-    public final ResponseEntity<ErrorDetails> handle(ExpiredTokenException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.UNAUTHORIZED);
-    }
-
-    @ExceptionHandler(GymNotFoundException.class)
-    public final ResponseEntity<ErrorDetails> handle(GymNotFoundException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(InvalidPasswordException.class)
-    public final ResponseEntity<ErrorDetails> handle(InvalidPasswordException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(InvalidReservationException.class)
-    public final ResponseEntity<ErrorDetails> handle(InvalidReservationException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(InvalidSaleException.class)
-    public final ResponseEntity<ErrorDetails> handle(InvalidSaleException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(InvalidEvent.class)
-    public final ResponseEntity<ErrorDetails> handle(InvalidEvent ex,
-                                                     WebRequest request) {
-        logger.info(ex.getMessage());
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(NotAllowedException.class)
-    public final ResponseEntity<ErrorDetails> handle(NotAllowedException ex,
+    @ExceptionHandler(MethodNotAllowedException.class)
+    public final ResponseEntity<ErrorDetails> handle(MethodNotAllowedException ex,
                                                      WebRequest request) {
         ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
                 request.getDescription(false));
         return new ResponseEntity<>(errorDetails, HttpStatus.METHOD_NOT_ALLOWED);
     }
 
-    @ExceptionHandler(RoleNotFoundException.class)
-    public final ResponseEntity<ErrorDetails> handle(RoleNotFoundException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(SalesIsNotCompletedException.class)
-    public final ResponseEntity<ErrorDetails> handle(SalesIsNotCompletedException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(TimesOffNotFound.class)
-    public final ResponseEntity<ErrorDetails> handle(TimesOffNotFound ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(TokenNotFoundException.class)
-    public final ResponseEntity<ErrorDetails> handle(TokenNotFoundException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(TrainingSessionNotFound.class)
-    public final ResponseEntity<ErrorDetails> handle(TrainingSessionNotFound ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(UnAuthorizedException.class)
-    public final ResponseEntity<ErrorDetails> handle(UnAuthorizedException ex,
+    @ExceptionHandler(UnauthorizedException.class)
+    public final ResponseEntity<ErrorDetails> handle(UnauthorizedException ex,
                                                      WebRequest request) {
         ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
                 request.getDescription(false));
         return new ResponseEntity<>(errorDetails, HttpStatus.UNAUTHORIZED);
     }
 
-    @ExceptionHandler(UserIsNotVerified.class)
-    public final ResponseEntity<ErrorDetails> handle(UserIsNotVerified ex,
+    @ExceptionHandler(BadRequestException.class)
+    public final ResponseEntity<ErrorDetails> handle(BadRequestException ex,
                                                      WebRequest request) {
         ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
                 request.getDescription(false));
         return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
     }
 
-    @ExceptionHandler(UserIsVerified.class)
-    public final ResponseEntity<ErrorDetails> handle(UserIsVerified ex,
+    @ExceptionHandler(ConflictException.class)
+    public final ResponseEntity<ErrorDetails> handle(ConflictException ex,
                                                      WebRequest request) {
         ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
                 request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(UserNotFoundException.class)
-    public final ResponseEntity<ErrorDetails> handle(UserNotFoundException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(UserRegistrationException.class)
-    public final ResponseEntity<ErrorDetails> handle(UserRegistrationException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-
-    @ExceptionHandler(InvalidTimeException.class)
-    public final ResponseEntity<ErrorDetails> handle(InvalidTimeException ex,
-                                                     WebRequest request) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), ex.getMessage(),
-                request.getDescription(false));
-        return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
+        return new ResponseEntity<>(errorDetails, HttpStatus.CONFLICT);
     }
 
 }